Added screenshotting functionality to offscreen renderer
This introduces a method to CefBrowser which is supposed to return a screenshot of the currently displayed browser window in the convenient form of a BufferedImage. CefBrowserOsr (offscreen rendering) implements this functionality by directly grabbing pixel data from the OpenGL rendering path. CefBrowserWr (windowed rendering) does not implement it because there is no clear way to access pixel data in case of windowed rendering.