Added possibility for HiDPI scaling correction
This is necessary in offscreen rendering situations if the result of the offscreen-rendered browser is to be displayed on a HiDPI display (with a non-1:1 mapping between physical and logical pixels). In such situations, JCEF would render things at the wrong scaling. With HiDPI scaling correction, one can fix the scaling (although the application is responsible for supplying a browser with the necessary info about the scaling factor).
In addition, while HiDPI works out of the box in windowed rendering mode on MacOS, it does not do so on Windows, hence this patch also enables HiDPI scaling on Windows in windowed and offscreen mode. There's the caveat however that in order to even use HiDPI on Windows, a Java 9+ runtime is needed (the Java 8 Windows runtime does not correctly support it).
This patch covers and was tested with MacOS and Windows. Linux should theoretically work much like Windows (if it doesn't even work out of the box), but as HiDPI support on Linux is pretty spotty still, it was not tested at all.