Update deps: JOGL 2.4-b1506, gluegen 2.4-b937
This fixes a longstanding JOGL bug which effectively terminates the application when using newer JDKs (1.8.0.262 and newer are known to cause problems). 
Unfortunately there is no 2.4 release of JOGL and gluegen yet, only beta versions (since 2015!), so we'll have to update to the newest one of those.

The original bug is tracked here: https://jogamp.org/bugzilla/show_bug.cgi?id=1370