Fixed rounding problem with OSR screenshotting
On some window sizes with "odd" scaling factors like 1.25, the calculation of the actual physical resolution can have the need to round values (for example 950px x 1,25 = 1187,5px) in order to end up with full integers again. Through experimentation it was found that the correct rounding algorithm is "Math.ceil", and not "Math.floor" which was implicitly done by casting to int until now. This error can cause slightly too small buffers when taking screenshots of windows with sizes that end up at "odd" widths or heights.